refactor: begin creating new Scott County parameters. move haart risk reduction. separate dx behavior change for injection/sex. race-specific needle exchange includes discontinuation only
the haart yml file was only used for a scaling factor that should have been in the calibration file. Behavior change may be specified in data based on type of risk act, so the code now reflects that. The logic of how the syringe services program now only calls for agents to leave the program at a low rate; the enrollment is determined by a cap.