fix: make print_components treatment bool work with PrEP and pca
print_components required a bool for whether it was a pca intervention in order to decide whether to count treatment components. This commit fixes that behavior, allowing treatment components to be printed in prep or pca interventions. This also fixes the definition of "intervention_ever" to only refer to randomized trial design"